requirements.txt
google-api-python-client>=2.100.0
together>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0